♻️(back) split chat views module and use shared constants
Split the monolithic chat views module into focused submodules
(conversations, attachments, projects, file streaming, filters, shared
helpers) while preserving the public import surface so existing callers
are unaffected.

Replace duplicated MIME-type and file-mode string literals across the
backend with shared constants and the existing upload-mode enums.
