+++ CHANGELOG.md
@@ -11,8 +11,12 @@ and this project adheres to
 ### Added
 
 - ✨(back) add model fallback mechanism
- - ✨(back) add celery for running background tasks
+- ✨(back) add celery for running background tasks
 - 🧱(helm) add celery worker and beat deployments
+
+### Changed
+
+- ♻️(back) split chat views and replace hardcoded strings with constants
  
 ## [0.0.19] - 2026-06-24
