+++ src/backend/chat/tests/conftest.py
@@ -14,6 +14,7 @@ import pytest
 import core.urls
 
 import chat.views
+import chat.views.conversations
 import conversations.urls
 from chat.agents.summarize import SummarizationAgent
 from chat.clients.pydantic_ai import AIAgentService
@@ -60,7 +61,9 @@ def mock_ai_agent_service_fixture():
             stack.enter_context(
                 patch("chat.clients.pydantic_ai.AIAgentService", new=AIAgentServiceMock)
             )
-            stack.enter_context(patch("chat.views.AIAgentService", new=AIAgentServiceMock))
+            stack.enter_context(
+                patch("chat.views.conversations.AIAgentService", new=AIAgentServiceMock)
+            )
             yield
 
     yield _mock_service
@@ -121,6 +124,10 @@ def fixture_oidc_refresh_token_enabled(settings):
     settings.OIDC_STORE_REFRESH_TOKEN = True
 
     # force view reload
+    # chat.views.conversations must be reloaded first: the conditional_refresh_oidc_token
+    # decorator is evaluated at import time there, and reloading the chat.views package
+    # alone would not re-execute the submodule.
+    reload(chat.views.conversations)
     reload(chat.views)
     reload(core.urls)
     reload(conversations.urls)
@@ -132,6 +139,7 @@ def fixture_oidc_refresh_token_enabled(settings):
     settings.OIDC_STORE_REFRESH_TOKEN = __initial_value
 
     # force view reload
+    reload(chat.views.conversations)
     reload(chat.views)
     reload(core.urls)
     reload(conversations.urls)
