+++ src/backend/chat/views/files.py
@@ -1,0 +1,79 @@
+"""Temporary-URL file streaming endpoint."""
+
+import logging
+
+from django.core.cache import cache
+from django.core.files.storage import default_storage
+from django.http import Http404, StreamingHttpResponse
+
+import magic
+from rest_framework.throttling import ScopedRateThrottle
+from rest_framework.views import APIView
+
+logger = logging.getLogger(__name__)
+
+
+class FileStreamView(APIView):
+    """
+    Stream file content for temporary access URLs.
+
+    This view is used by LLMs to access file content when they cannot directly
+    access S3. A temporary key is stored in cache and validated before serving
+    the file.
+
+    Security:
+    - Temporary key expires after FILE_BACKEND_TEMPORARY_URL_EXPIRATION seconds
+      (default: 180 seconds / 3 minutes)
+    - No authentication required (key is single-use temporary token)
+    - Key is generated using secure random tokens
+    """
+
+    permission_classes = []  # No authentication needed for temporary keys
+    throttle_classes = [ScopedRateThrottle]
+    throttle_scope = "file-stream"
+
+    def get(self, request, temporary_key):
+        """
+        Stream file content using a temporary access key.
+
+        Args:
+            temporary_key: The temporary key generated by generate_temporary_url()
+
+        Returns:
+            StreamingHttpResponse with file content
+        """
+        # Retrieve the S3 key from cache using the temporary key
+        cache_key = f"file_access:{temporary_key}"
+        s3_key = cache.get(cache_key)
+
+        if not s3_key:
+            logger.warning("Temporary file access key not found or expired: %s", temporary_key)
+            raise Http404("File access key expired or invalid")
+
+        cache.delete(cache_key)  # Delete the key from cache to prevent reuse
+
+        logger.info("Serving file via temporary key: %s", s3_key)
+
+        try:
+            # Open the file from S3
+            file_obj = default_storage.open(s3_key, "rb")
+
+            # Detect MIME type for proper content-type header
+            mime_detector = magic.Magic(mime=True)
+            file_content = file_obj.read(2048)
+            file_obj.seek(0)
+            content_type = mime_detector.from_buffer(file_content)
+
+            # Extract filename from S3 key (last part after /)
+            filename = s3_key.split("/")[-1]
+
+            # Stream the file content
+            response = StreamingHttpResponse(
+                file_obj,
+                content_type=content_type,
+            )
+            response["Content-Disposition"] = f'inline; filename="{filename}"'
+            return response
+        except Exception as exc:
+            logger.exception("Failed to serve file via temporary key: %s", temporary_key)
+            raise Http404("Failed to retrieve file") from exc
